+++ programs/quarry-operator/src/instructions/mod.rs
@@ -1,5 +1,7 @@
+pub mod create_operator;
 pub mod delegate_create_quarry;
 pub mod delegate_create_quarry_v2;
 
+pub use create_operator::*;
 pub use delegate_create_quarry::*;
 pub use delegate_create_quarry_v2::*;
